backend/requirements.txt
psycopg2-binary
python-dotenv
fastapi
uvicorn[standard]
sqlmodel
aiosqlite
requests
pytest
httpx
pytest-asyncio
